+++ .github/workflows/github-pages.yml
@@ -5,7 +5,9 @@ on:
   # Runs on pushes targeting the default branch
   push:
     # Run on changes in the master branch
-    #branches: [master]:
+    branches:
+      - master
+      - docusaurus
 
     # Run on changes in the docs folder
     paths:
@@ -36,6 +38,9 @@ jobs:
       name: github-pages
       url: ${{ steps.deployment.outputs.page_url }}
     runs-on: ubuntu-latest
+
+    # execute only in Louisvdw repository
+    if: github.repository_owner == 'Louisvdw'
 
     #defaults:
     #  run:
