+++ .github/workflows/release.yml
@@ -11,18 +11,20 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v3
+
       - name: build release archive
         run: |
+          find . -type f -name "*.py" -exec chmod +x {} \;
           find . -type f -name "*.sh" -exec chmod +x {} \;
+          find . -type f -name "run" -exec chmod +x {} \;
           tar -czvf venus-data.tar.gz \
             --mode='a+rwX' \
-            --exclude battery_template.py \
             --exclude __pycache__ \
-            --exclude restartservice.sh \
-            --exclude revov.py \
-            --exclude test_max17853.py \
-            conf/serial-starter.d \
+            --exclude bms/battery_template.py \
+            --exclude bms/revov.py \
+            --exclude bms/test_max17853.py \
             etc/dbus-serialbattery/
+
       - name: Release
         uses: softprops/action-gh-release@v1
         with:
