+++ .gitignore
@@ -151,6 +151,9 @@ cython_debug/
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
 
+# VS Code
+.vscode/
+
 # Custom for this repo
 venus-data.tar.gz
 BMS-trials
@@ -158,6 +161,7 @@ BMS-trials
 .project
 .pydevproject
 *.prefs
+etc/dbus-serialbattery/config.ini
 
 # Local Clone of velib_python
 velib_python
