--- conf/serial-starter.d
@@ -1,3 +1,0 @@
-service sbattery        dbus-serialbattery
-alias default gps:vedirect:sbattery
-alias rs485 cgwacs:fzsonick:imt:modbus:sbattery
