+++ etc/dbus-serialbattery/README.md
@@ -1,2 +1,5 @@
 # dbus-serialbattery
-See [README on GitHub](https://github.com/Louisvdw/dbus-serialbattery/blob/master/README.md)
+
+* See [README](https://github.com/Louisvdw/dbus-serialbattery/blob/master/README.md) on GitHub
+
+* See [Documentation](https://louisvdw.github.io/dbus-serialbattery/) on GitHub Pages
