+++ etc/dbus-serialbattery/bms/jkbms.py
@@ -3,6 +3,7 @@ from battery import Battery, Cell
 from utils import is_bit_set, read_serial_data, logger
 import utils
 from struct import unpack_from
+from re import sub
 
 
 class Jkbms(Battery):
@@ -18,6 +19,9 @@ class Jkbms(Battery):
     command_status = b"\x4E\x57\x00\x13\x00\x00\x00\x00\x06\x03\x00\x00\x00\x00\x00\x00\x68\x00\x00\x01\x29"
 
     def test_connection(self):
+        # call a function that will connect to the battery, send a command and retrieve the result.
+        # The result or call should be unique to this BMS. Battery name or version, etc.
+        # Return True if success, False for failure
         try:
             return self.read_status_data()
         except Exception as err:
@@ -28,8 +32,6 @@ class Jkbms(Battery):
         # After successful  connection get_settings will be call to set up the battery.
         # Set the current limits, populate cell count, etc
         # Return True if success, False for failure
-        self.max_battery_charge_current = utils.MAX_BATTERY_CHARGE_CURRENT
-        self.max_battery_discharge_current = utils.MAX_BATTERY_DISCHARGE_CURRENT
         self.max_battery_voltage = utils.MAX_CELL_VOLTAGE * self.cell_count
         self.min_battery_voltage = utils.MIN_CELL_VOLTAGE * self.cell_count
 
@@ -81,17 +83,19 @@ class Jkbms(Battery):
                     unpack_from(">xH", celldata, c * 3 + 1)[0] / 1000
                 )
 
+        # MOSFET temperature
+        offset = cellbyte_count + 3
+        temp_mos = unpack_from(">H", self.get_data(status_data, b"\x80", offset, 2))[0]
+        self.to_temp(0, temp_mos if temp_mos < 99 else (100 - temp_mos))
+
+        # Temperature sensors
         offset = cellbyte_count + 6
         temp1 = unpack_from(">H", self.get_data(status_data, b"\x81", offset, 2))[0]
+
         offset = cellbyte_count + 9
         temp2 = unpack_from(">H", self.get_data(status_data, b"\x82", offset, 2))[0]
         self.to_temp(1, temp1 if temp1 < 99 else (100 - temp1))
         self.to_temp(2, temp2 if temp2 < 99 else (100 - temp2))
-
-        # MOSFET temperature
-        offset = cellbyte_count + 3
-        temp_mos = unpack_from(">H", self.get_data(status_data, b"\x80", offset, 2))[0]
-        self.to_temp("mos", temp_mos if temp_mos < 99 else (100 - temp_mos))
 
         offset = cellbyte_count + 12
         voltage = unpack_from(">H", self.get_data(status_data, b"\x83", offset, 2))[0]
@@ -103,6 +107,18 @@ class Jkbms(Battery):
             current / -100
             if current < self.CURRENT_ZERO_CONSTANT
             else (current - self.CURRENT_ZERO_CONSTANT) / 100
+        )
+
+        # Continued discharge current
+        offset = cellbyte_count + 66
+        self.max_battery_discharge_current = float(
+            unpack_from(">H", self.get_data(status_data, b"\x97", offset, 2))[0]
+        )
+
+        # Continued charge current
+        offset = cellbyte_count + 72
+        self.max_battery_charge_current = float(
+            unpack_from(">H", self.get_data(status_data, b"\x99", offset, 2))[0]
         )
 
         offset = cellbyte_count + 18
@@ -124,56 +140,157 @@ class Jkbms(Battery):
         self.to_protection_bits(
             unpack_from(">H", self.get_data(status_data, b"\x8B", offset, 2))[0]
         )
+
         offset = cellbyte_count + 36
         self.to_fet_bits(
             unpack_from(">H", self.get_data(status_data, b"\x8C", offset, 2))[0]
         )
 
+        offset = cellbyte_count + 84
+        self.to_balance_bits(
+            unpack_from(">B", self.get_data(status_data, b"\x9D", offset, 1))[0]
+        )
+
+        # User Private Data filed in APP
         offset = cellbyte_count + 155
-        self.production = unpack_from(
-            ">8s", self.get_data(status_data, b"\xB4", offset, 8)
-        )[0].decode()
+        self.production = sub(
+            " +",
+            " ",
+            (
+                unpack_from(">8s", self.get_data(status_data, b"\xB4", offset, 8))[0]
+                .decode()
+                .replace("\x00", " ")
+                .strip()
+            ),
+        )
+        self.production = self.production if self.production != "Input Us" else None
+
         offset = cellbyte_count + 174
         self.version = unpack_from(
             ">15s", self.get_data(status_data, b"\xB7", offset, 15)
         )[0].decode()
 
+        offset = cellbyte_count + 197
+        self.unique_identifier = sub(
+            " +",
+            " ",
+            (
+                unpack_from(">24s", self.get_data(status_data, b"\xBA", offset, 24))[0]
+                .decode()
+                .replace("\x00", " ")
+                .replace("Input Userda", "")
+                .strip()
+            ),
+        )
+
+        # show wich cells are balancing
+        if self.get_min_cell() is not None and self.get_max_cell() is not None:
+            for c in range(self.cell_count):
+                if self.balancing and (
+                    self.get_min_cell() == c or self.get_max_cell() == c
+                ):
+                    self.cells[c].balance = True
+                else:
+                    self.cells[c].balance = False
+
         # logger.info(self.hardware_version)
         return True
 
     def to_fet_bits(self, byte_data):
-        tmp = bin(byte_data)[2:].rjust(2, utils.zero_char)
-        self.charge_fet = is_bit_set(tmp[1])
-        self.discharge_fet = is_bit_set(tmp[0])
+        tmp = bin(byte_data)[2:].rjust(3, utils.zero_char)
+        self.charge_fet = is_bit_set(tmp[2])
+        self.discharge_fet = is_bit_set(tmp[1])
+        self.balancing = is_bit_set(tmp[0])
+
+    def to_balance_bits(self, byte_data):
+        tmp = bin(byte_data)[2:]
+        self.balance_fet = is_bit_set(tmp)
+
+    def get_balancing(self):
+        return 1 if self.balancing else 0
+
+    def get_min_cell(self):
+        min_voltage = 9999
+        min_cell = None
+        for c in range(min(len(self.cells), self.cell_count)):
+            if (
+                self.cells[c].voltage is not None
+                and min_voltage > self.cells[c].voltage
+            ):
+                min_voltage = self.cells[c].voltage
+                min_cell = c
+        return min_cell
+
+    def get_max_cell(self):
+        max_voltage = 0
+        max_cell = None
+        for c in range(min(len(self.cells), self.cell_count)):
+            if (
+                self.cells[c].voltage is not None
+                and max_voltage < self.cells[c].voltage
+            ):
+                max_voltage = self.cells[c].voltage
+                max_cell = c
+        return max_cell
 
     def to_protection_bits(self, byte_data):
+        """
+        Bit 0: Low capacity alarm: 1 warning only, 0 nomal -> OK
+        Bit 1: MOS tube overtemperature alarm: 1 alarm, 0 nomal -> OK
+        Bit 2: Charge over voltage alarm: 1 alarm, 0 nomal -> OK
+        Bit 3: Discharge undervoltage alarm: 1 alarm, 0 nomal -> OK
+        Bit 4: Battery overtemperature alarm: 1 alarm, 0 nomal -> OK
+        Bit 5: Charge overcurrent alarm: 1 alarm, 0 nomal -> OK
+        Bit 6: discharge over current alarm: 1 alarm, 0 nomal -> OK
+        Bit 7: core differential pressure alarm: 1 alarm, 0 nomal -> OK
+        Bit 8: overtemperature alarm in the battery box: 1 alarm, 0 nomal -> OK
+        Bit 9: Battery low temperature alarm: 1 alarm, 0 nomal -> OK
+        Bit 10: Unit overvoltage: 1 alarm, 0 nomal -> OK
+        Bit 11: Unit undervoltage: 1 alarm, 0 nomal -> OK
+        Bit 12:309_A protection: 1 alarm, 0 nomal
+        Bit 13:309_B protection: 1 alarm, 0 nomal
+        """
         pos = 13
         tmp = bin(byte_data)[15 - pos :].rjust(pos + 1, utils.zero_char)
         # logger.debug(tmp)
+
+        # low capacity alarm
         self.protection.soc_low = 2 if is_bit_set(tmp[pos - 0]) else 0
-        self.protection.set_IC_inspection = (
-            2 if is_bit_set(tmp[pos - 1]) else 0
-        )  # BMS over temp
+        # MOSFET temperature alarm
+        self.protection.temp_high_internal = 2 if is_bit_set(tmp[pos - 1]) else 0
+        # charge over voltage alarm
         self.protection.voltage_high = 2 if is_bit_set(tmp[pos - 2]) else 0
+        # discharge under voltage alarm
         self.protection.voltage_low = 2 if is_bit_set(tmp[pos - 3]) else 0
+        # charge overcurrent alarm
         self.protection.current_over = 1 if is_bit_set(tmp[pos - 5]) else 0
+        # discharge over current alarm
         self.protection.current_under = 1 if is_bit_set(tmp[pos - 6]) else 0
+        # core differential pressure alarm OR unit overvoltage alarm
         self.protection.cell_imbalance = (
             2 if is_bit_set(tmp[pos - 7]) else 1 if is_bit_set(tmp[pos - 10]) else 0
         )
-        self.protection.voltage_cell_low = 2 if is_bit_set(tmp[pos - 11]) else 0
-        # there is just a BMS and Battery temp alarm (not high/low)
-        self.protection.temp_high_charge = (
+        # unit undervoltage alarm
+        self.protection.voltage_cell_low = 1 if is_bit_set(tmp[pos - 11]) else 0
+        # battery overtemperature alarm OR overtemperature alarm in the battery box
+        alarm_temp_high = (
             1 if is_bit_set(tmp[pos - 4]) or is_bit_set(tmp[pos - 8]) else 0
+        )
+        # battery low temperature alarm
+        alarm_temp_low = 1 if is_bit_set(tmp[pos - 9]) else 0
+        # check if low/high temp alarm arise during charging
+        self.protection.temp_high_charge = (
+            1 if self.current > 0 and alarm_temp_high == 1 else 0
         )
         self.protection.temp_low_charge = (
-            1 if is_bit_set(tmp[pos - 4]) or is_bit_set(tmp[pos - 8]) else 0
+            1 if self.current > 0 and alarm_temp_low == 1 else 0
         )
+        # check if low/high temp alarm arise during discharging
         self.protection.temp_high_discharge = (
-            1 if is_bit_set(tmp[pos - 4]) or is_bit_set(tmp[pos - 8]) else 0
+            1 if self.current <= 0 and alarm_temp_high == 1 else 0
         )
         self.protection.temp_low_discharge = (
-            1 if is_bit_set(tmp[pos - 4]) or is_bit_set(tmp[pos - 8]) else 0
+            1 if self.current <= 0 and alarm_temp_low == 1 else 0
         )
 
     def read_serial_data_jkbms(self, command: str) -> bool:
@@ -200,7 +317,7 @@ class Jkbms(Battery):
         s = sum(data[0:-4])
 
         if start == 0x4E57 and end == 0x68 and s == crc_lo:
-            return data[10 : length - 19]
+            return data[10 : length - 7]
         elif s != crc_lo:
             logger.error(
                 "CRC checksum mismatch: Expected 0x%04x, Got 0x%04x" % (crc_lo, s)
