+++ etc/dbus-serialbattery/dbus-serialbattery.py
@@ -4,7 +4,8 @@ from typing import Union
 
 from time import sleep
 from dbus.mainloop.glib import DBusGMainLoop
-from threading import Thread
+
+# from threading import Thread  ## removed with https://github.com/Louisvdw/dbus-serialbattery/pull/582
 import sys
 
 if sys.version_info.major == 2:
@@ -19,31 +20,35 @@ from dbushelper import DbusHelper
 from utils import logger
 import utils
 from battery import Battery
-from lltjbd import LltJbd
-from daly import Daly
-from ant import Ant
-from jkbms import Jkbms
-from seplos import Seplos
 
-# from sinowealth import Sinowealth
-from renogy import Renogy
-from ecs import Ecs
-from lifepower import Lifepower
-from hlpdatabms4s import HLPdataBMS4S
+# import battery classes
+from bms.daly import Daly
+from bms.ecs import Ecs
+from bms.hlpdatabms4s import HLPdataBMS4S
+from bms.jkbms import Jkbms
+from bms.lifepower import Lifepower
+from bms.lltjbd import LltJbd
+from bms.renogy import Renogy
+from bms.seplos import Seplos
+
+# from bms.ant import Ant
+# from bms.mnb import MNB
+# from bms.sinowealth import Sinowealth
 
 supported_bms_types = [
-    {'bms': HLPdataBMS4S, "baud": 9600},
-    {"bms": LltJbd, "baud": 9600},
-    {"bms": Ant, "baud": 19200},
     {"bms": Daly, "baud": 9600, "address": b"\x40"},
     {"bms": Daly, "baud": 9600, "address": b"\x80"},
+    {"bms": Ecs, "baud": 19200},
+    {"bms": HLPdataBMS4S, "baud": 9600},
     {"bms": Jkbms, "baud": 115200},
-    #    {"bms" : Sinowealth},
     {"bms": Lifepower, "baud": 9600},
+    {"bms": LltJbd, "baud": 9600},
     {"bms": Renogy, "baud": 9600, "address": b"\x30"},
     {"bms": Renogy, "baud": 9600, "address": b"\xF7"},
-    {"bms": Ecs, "baud": 19200},
     {"bms": Seplos, "baud": 19200},
+    # {"bms": Ant, "baud": 19200},
+    # {"bms": MNB, "baud": 9600},
+    # {"bms": Sinowealth},
 ]
 expected_bms_types = [
     battery_type
@@ -51,16 +56,13 @@ expected_bms_types = [
     if battery_type["bms"].__name__ == utils.BMS_TYPE or utils.BMS_TYPE == ""
 ]
 
+logger.info("")
 logger.info("Starting dbus-serialbattery")
 
 
 def main():
     def poll_battery(loop):
-        # Run in separate thread. Pass in the mainloop so the thread can kill us if there is an exception.
-        poller = Thread(target=lambda: helper.publish_battery(loop))
-        # Thread will die with us if deamon
-        poller.daemon = True
-        poller.start()
+        helper.publish_battery(loop)
         return True
 
     def get_battery(_port) -> Union[Battery, None]:
@@ -102,6 +104,7 @@ def main():
     port = get_port()
     battery: Battery = get_battery(port)
 
+    # exit if no battery could be found
     if battery is None:
         logger.error("ERROR >>> No battery connection at " + port)
         sys.exit(1)
