--- etc/dbus-serialbattery/disabledriver.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-set -x
-rm -f /data/conf/serial-starter.d
