--- etc/dbus-serialbattery/installlocal.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-set -x
-tar -zxf ./venus-data.tar.gz -C /data
-sh /data/etc/dbus-serialbattery/reinstalllocal.sh
