--- etc/dbus-serialbattery/restartservice.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-svc -d -u /service/dbus-serialbattery
