+++ etc/dbus-serialbattery/utils.py
@@ -17,7 +17,7 @@ logger.setLevel(logging.INFO)
 
 config = configparser.ConfigParser()
 path = Path(__file__).parents[0]
-default_config_file_path = path.joinpath("default_config.ini").absolute().__str__()
+default_config_file_path = path.joinpath("config.default.ini").absolute().__str__()
 custom_config_file_path = path.joinpath("config.ini").absolute().__str__()
 config.read([default_config_file_path, custom_config_file_path])
 
@@ -35,34 +35,97 @@ def _get_list_from_config(
 # if not specified: baud = 9600
 
 # Constants - Need to dynamically get them in future
-DRIVER_VERSION = 0.14
-DRIVER_SUBVERSION = ".3"
+DRIVER_VERSION = "1.0"
+DRIVER_SUBVERSION = ".0 (20230505)"
 zero_char = chr(48)
 degree_sign = "\N{DEGREE SIGN}"
 
-# Choose the mode for voltage / current limitations (True / False)
-# False is a Step mode. This is the default with limitations on hard boundary steps
-# True "Linear"    # New linear limitations by WaldemarFech for smoother values
-LINEAR_LIMITATION_ENABLE = "True" == config["DEFAULT"]["LINEAR_LIMITATION_ENABLE"]
-
-# battery Current limits
+# --------- Battery Current limits ---------
 MAX_BATTERY_CHARGE_CURRENT = float(config["DEFAULT"]["MAX_BATTERY_CHARGE_CURRENT"])
 MAX_BATTERY_DISCHARGE_CURRENT = float(
     config["DEFAULT"]["MAX_BATTERY_DISCHARGE_CURRENT"]
 )
 
-# -------- Cell Voltage limitation ---------
-# Description:
-# Maximal charge / discharge current will be in-/decreased depending on min- and max-cell-voltages
-# Example: 18cells * 3.55V/cell = 63.9V max charge voltage. 18 * 2.7V = 48,6V min discharge voltage
-#          ... but the (dis)charge current will be (in-/)decreased, if even ONE SINGLE BATTERY CELL reaches the limits
+# --------- Cell Voltages ---------
+# Description: Cell min/max voltages which are used to calculate the min/max battery voltage
+# Example: 16 cells * 3.45V/cell = 55.2V max charge voltage. 16 cells * 2.90V = 46.4V min discharge voltage
+MIN_CELL_VOLTAGE = float(config["DEFAULT"]["MIN_CELL_VOLTAGE"])
+MAX_CELL_VOLTAGE = float(config["DEFAULT"]["MAX_CELL_VOLTAGE"])
+# Max voltage can seen as absorption voltage
+FLOAT_CELL_VOLTAGE = float(config["DEFAULT"]["FLOAT_CELL_VOLTAGE"])
+
+# --------- BMS disconnect behaviour ---------
+# Description: Block charge and discharge when the communication to the BMS is lost. If you are removing the
+#              BMS on purpose, then you have to restart the driver/system to reset the block.
+# False: Charge and discharge is not blocked on BMS communication loss
+# True: Charge and discharge is blocked on BMS communication loss, it's unblocked when connection is established
+#       again or the driver/system is restarted
+BLOCK_ON_DISCONNECT = "True" == config["DEFAULT"]["BLOCK_ON_DISCONNECT"]
+
+# --------- Charge mode ---------
+# Choose the mode for voltage / current limitations (True / False)
+# False is a step mode. This is the default with limitations on hard boundary steps
+# True is a linear mode. For CCL and DCL the values between the steps are calculated for smoother values (by WaldemarFech)
+#                        For CVL max battery voltage is calculated dynamically in order that the max cell voltage is not exceeded
+LINEAR_LIMITATION_ENABLE = "True" == config["DEFAULT"]["LINEAR_LIMITATION_ENABLE"]
+
+# Specify in seconds how often the penalty should be recalculated
+LINEAR_RECALCULATION_EVERY = int(config["DEFAULT"]["LINEAR_RECALCULATION_EVERY"])
+# Specify in percent when the linear values should be recalculated immediately
+# Example: 5 for a immediate change, when the value changes by more than 5%
+LINEAR_RECALCULATION_ON_PERC_CHANGE = int(
+    config["DEFAULT"]["LINEAR_RECALCULATION_ON_PERC_CHANGE"]
+)
+
+
+# --------- Charge Voltage limitation (affecting CVL) ---------
+# Description: Limit max charging voltage (MAX_CELL_VOLTAGE * cell count), switch from max voltage to float voltage (FLOAT_CELL_VOLTAGE * cell count) and back
+#     Step mode: After max voltage is reached for MAX_VOLTAGE_TIME_SEC it switches to float voltage. After SoC is below SOC_LEVEL_TO_RESET_VOLTAGE_LIMIT it
+#                switches back to max voltage.
+#     Linear mode: After max voltage is reachend and cell voltage difference is smaller or equal to CELL_VOLTAGE_DIFF_KEEP_MAX_VOLTAGE_UNTIL it switches to
+#                  float voltage after 300 (fixed) additional seconds. After cell voltage difference is greater or equal to CELL_VOLTAGE_DIFF_TO_RESET_VOLTAGE_LIMIT
+#                  it switches back to max voltage.
+# Example: The battery reached max voltage of 55.2V and hold it for 900 seconds, the the CVL is switched to float voltage of 53.6V to don't stress the batteries.
+#          Allow max voltage of 55.2V again, if SoC is once below 90%
+#          OR
+#          The battery reached max voltage of 55.2V and the max cell difference is 0.010V, then switch to float voltage of 53.6V after 300 additional seconds
+#          to don't stress the batteries. Allow max voltage of 55.2V again if max cell difference is above 0.050V
+# Charge voltage control management enable (True/False).
+CVCM_ENABLE = "True" == config["DEFAULT"]["CVCM_ENABLE"]
+
+# -- CVL reset based on cell voltage diff (linear mode)
+# Specify cell voltage diff where CVL limit is kept until diff is equal or lower
+CELL_VOLTAGE_DIFF_KEEP_MAX_VOLTAGE_UNTIL = float(
+    config["DEFAULT"]["CELL_VOLTAGE_DIFF_KEEP_MAX_VOLTAGE_UNTIL"]
+)
+# Specify cell voltage diff where CVL limit is reset to max voltage, if value get above
+CELL_VOLTAGE_DIFF_TO_RESET_VOLTAGE_LIMIT = float(
+    config["DEFAULT"]["CELL_VOLTAGE_DIFF_TO_RESET_VOLTAGE_LIMIT"]
+)
+
+# -- CVL Reset based on SoC option
+# Reset max voltage after
+MAX_VOLTAGE_TIME_SEC = float(config["DEFAULT"]["MAX_VOLTAGE_TIME_SEC"])
+# Specify SoC where CVL limit is reset to max voltage
+SOC_LEVEL_TO_RESET_VOLTAGE_LIMIT = float(
+    config["DEFAULT"]["SOC_LEVEL_TO_RESET_VOLTAGE_LIMIT"]
+)
+
+
+# --------- Cell Voltage Current limitation (affecting CCL/DCL) ---------
+# Description: Maximal charge / discharge current will be in-/decreased depending on min and max cell voltages
+# Example: 18 cells * 3.55V/cell = 63.9V max charge voltage
+#          18 cells * 2.70V/cell = 48.6V min discharge voltage
+#          But in reality not all cells reach the same voltage at the same time. The (dis)charge current
+#          will be (in-/)decreased, if even ONE SINGLE BATTERY CELL reaches the limits
 
 # Charge current control management referring to cell-voltage enable (True/False).
 CCCM_CV_ENABLE = "True" == config["DEFAULT"]["CCCM_CV_ENABLE"]
 # Discharge current control management referring to cell-voltage enable (True/False).
 DCCM_CV_ENABLE = "True" == config["DEFAULT"]["DCCM_CV_ENABLE"]
 
-# Set Steps to reduce battery current. The current will be changed linear between those steps
+# Set steps to reduce battery current
+# The current will be changed linear between those steps if LINEAR_LIMITATION_ENABLE is set to True
 CELL_VOLTAGES_WHILE_CHARGING = _get_list_from_config(
     "DEFAULT", "CELL_VOLTAGES_WHILE_CHARGING", lambda v: float(v)
 )
@@ -81,9 +144,9 @@ MAX_DISCHARGE_CURRENT_CV = _get_list_from_config(
     lambda v: MAX_BATTERY_DISCHARGE_CURRENT * float(v),
 )
 
-# -------- Temperature limitation ---------
-# Description:
-# Maximal charge / discharge current will be in-/decreased depending on temperature
+
+# --------- Temperature limitation (affecting CCL/DCL) ---------
+# Description: Maximal charge / discharge current will be in-/decreased depending on temperature
 # Example: The temperature limit will be monitored to control the currents. If there are two temperature senors,
 #          then the worst case will be calculated and the more secure lower current will be set.
 # Charge current control management referring to temperature enable (True/False).
@@ -91,7 +154,8 @@ CCCM_T_ENABLE = "True" == config["DEFAULT"]["CCCM_T_ENABLE"]
 # Charge current control management referring to temperature enable (True/False).
 DCCM_T_ENABLE = "True" == config["DEFAULT"]["DCCM_T_ENABLE"]
 
-# Set Steps to reduce battery current. The current will be changed linear between those steps
+# Set steps to reduce battery current
+# The current will be changed linear between those steps if LINEAR_LIMITATION_ENABLE is set to True
 TEMPERATURE_LIMITS_WHILE_CHARGING = _get_list_from_config(
     "DEFAULT", "TEMPERATURE_LIMITS_WHILE_CHARGING", lambda v: float(v)
 )
@@ -110,42 +174,22 @@ MAX_DISCHARGE_CURRENT_T = _get_list_from_config(
     lambda v: MAX_BATTERY_DISCHARGE_CURRENT * float(v),
 )
 
-# if the cell voltage reaches 3.55V, then reduce current battery-voltage by 0.01V
-# if the cell voltage goes over 3.6V, then the maximum penalty will not be exceeded
-# there will be a sum of all penalties for each cell, which exceeds the limits
-PENALTY_AT_CELL_VOLTAGE = _get_list_from_config(
-    "DEFAULT", "PENALTY_AT_CELL_VOLTAGE", lambda v: float(v)
-)
-PENALTY_BATTERY_VOLTAGE = _get_list_from_config(
-    "DEFAULT", "PENALTY_BATTERY_VOLTAGE", lambda v: float(v)
-)
 
-
-# -------- SOC limitation ---------
-# Description:
-# Maximal charge / discharge current will be increased / decreased depending on State of Charge, see CC_SOC_LIMIT1 etc.
-# The State of Charge (SoC) charge / discharge current will be in-/decreased depending on SOC.
-# Example: 16cells * 3.45V/cell = 55,2V max charge voltage. 16*2.9V = 46,4V min discharge voltage
-# Cell min/max voltages - used with the cell count to get the min/max battery voltage
-MIN_CELL_VOLTAGE = float(config["DEFAULT"]["MIN_CELL_VOLTAGE"])
-MAX_CELL_VOLTAGE = float(config["DEFAULT"]["MAX_CELL_VOLTAGE"])
-FLOAT_CELL_VOLTAGE = float(config["DEFAULT"]["FLOAT_CELL_VOLTAGE"])
-MAX_VOLTAGE_TIME_SEC = float(config["DEFAULT"]["MAX_VOLTAGE_TIME_SEC"])
-SOC_LEVEL_TO_RESET_VOLTAGE_LIMIT = float(
-    config["DEFAULT"]["SOC_LEVEL_TO_RESET_VOLTAGE_LIMIT"]
-)
-
+# --------- SOC limitation (affecting CCL/DCL) ---------
+# Description: Maximal charge / discharge current will be increased / decreased depending on State of Charge,
+#              see CC_SOC_LIMIT1 etc.
+# Example: The SoC limit will be monitored to control the currents.
 # Charge current control management enable (True/False).
 CCCM_SOC_ENABLE = "True" == config["DEFAULT"]["CCCM_SOC_ENABLE"]
 # Discharge current control management enable (True/False).
 DCCM_SOC_ENABLE = "True" == config["DEFAULT"]["DCCM_SOC_ENABLE"]
 
-# charge current soc limits
+# Charge current soc limits
 CC_SOC_LIMIT1 = float(config["DEFAULT"]["CC_SOC_LIMIT1"])
 CC_SOC_LIMIT2 = float(config["DEFAULT"]["CC_SOC_LIMIT2"])
 CC_SOC_LIMIT3 = float(config["DEFAULT"]["CC_SOC_LIMIT3"])
 
-# charge current limits
+# Charge current limits
 CC_CURRENT_LIMIT1 = MAX_BATTERY_CHARGE_CURRENT * float(
     config["DEFAULT"]["CC_CURRENT_LIMIT1_FRACTION"]
 )
@@ -156,12 +200,12 @@ CC_CURRENT_LIMIT3 = MAX_BATTERY_CHARGE_CURRENT * float(
     config["DEFAULT"]["CC_CURRENT_LIMIT3_FRACTION"]
 )
 
-# discharge current soc limits
+# Discharge current soc limits
 DC_SOC_LIMIT1 = float(config["DEFAULT"]["DC_SOC_LIMIT1"])
 DC_SOC_LIMIT2 = float(config["DEFAULT"]["DC_SOC_LIMIT2"])
 DC_SOC_LIMIT3 = float(config["DEFAULT"]["DC_SOC_LIMIT3"])
 
-# discharge current limits
+# Discharge current limits
 DC_CURRENT_LIMIT1 = MAX_BATTERY_DISCHARGE_CURRENT * float(
     config["DEFAULT"]["DC_CURRENT_LIMIT1_FRACTION"]
 )
@@ -172,57 +216,96 @@ DC_CURRENT_LIMIT3 = MAX_BATTERY_DISCHARGE_CURRENT * float(
     config["DEFAULT"]["DC_CURRENT_LIMIT3_FRACTION"]
 )
 
-# Charge voltage control management enable (True/False).
-CVCM_ENABLE = "True" == config["DEFAULT"]["CVCM_ENABLE"]
+
+# --------- Time-To-Go ---------
+# Description: Calculates the time to go shown in the GUI
+TIME_TO_GO_ENABLE = "True" == config["DEFAULT"]["TIME_TO_GO_ENABLE"]
+
+# --------- Time-To-Soc ---------
+# Description: Calculates the time to a specific SoC
+# Example: TIME_TO_SOC_POINTS = 50, 25, 15, 0
+#          6h 24m remaining until 50% SoC
+#          17h 36m remaining until 25% SoC
+#          22h 5m remaining until 15% SoC
+#          28h 48m remaining until 0% SoC
+# Set of SoC percentages to report on dbus and MQTT. The more you specify the more it will impact system performance.
+# [Valid values 0-100, comma separated list. More that 20 intervals are not recommended]
+# Example: TIME_TO_SOC_POINTS = 100, 95, 90, 85, 75, 50, 25, 20, 10, 0
+# Leave empty to disable
+TIME_TO_SOC_POINTS = _get_list_from_config(
+    "DEFAULT", "TIME_TO_SOC_POINTS", lambda v: int(v)
+)
+# Specify TimeToSoc value type [Valid values 1, 2, 3]
+# 1 Seconds
+# 2 Time string <days>d <hours>h <minutes>m <seconds>s
+# 3 Both seconds and time string "<seconds> [<days>d <hours>h <minutes>m <seconds>s]"
+TIME_TO_SOC_VALUE_TYPE = int(config["DEFAULT"]["TIME_TO_SOC_VALUE_TYPE"])
+# Specify in seconds how often the TimeToSoc should be recalculated
+# Minimum are 5 seconds to prevent CPU overload
+TIME_TO_SOC_RECALCULATE_EVERY = (
+    int(config["DEFAULT"]["TIME_TO_SOC_RECALCULATE_EVERY"])
+    if int(config["DEFAULT"]["TIME_TO_SOC_RECALCULATE_EVERY"]) > 5
+    else 5
+)
+# Include TimeToSoC points when moving away from the SoC point [Valid values True, False]
+# These will be as negative time. Disabling this improves performance slightly
+TIME_TO_SOC_INC_FROM = "True" == config["DEFAULT"]["TIME_TO_SOC_INC_FROM"]
+
+
+# --------- Additional settings ---------
+# Specify only one BMS type to load else leave empty to try to load all availabe
+# LltJbd, Ant, Daly, Daly, Jkbms, Lifepower, Renogy, Renogy, Ecs
+BMS_TYPE = config["DEFAULT"]["BMS_TYPE"]
+
+# Publish the config settings to the dbus path "/Info/Config/"
+PUBLISH_CONFIG_VALUES = int(config["DEFAULT"]["PUBLISH_CONFIG_VALUES"])
+
+# Select the format of cell data presented on dbus [Valid values 0,1,2,3]
+# 0 Do not publish all the cells (only the min/max cell data as used by the default GX)
+# 1 Format: /Voltages/Cell (also available for display on Remote Console)
+# 2 Format: /Cell/#/Volts
+# 3 Both formats 1 and 2
+BATTERY_CELL_DATA_FORMAT = int(config["DEFAULT"]["BATTERY_CELL_DATA_FORMAT"])
 
 # Simulate Midpoint graph (True/False).
 MIDPOINT_ENABLE = "True" == config["DEFAULT"]["MIDPOINT_ENABLE"]
 
-# soc low levels
+# Battery temperature
+# Specifiy how the battery temperature is assembled
+# 0 Get mean of temp sensor 1 and temp sensor 2
+# 1 Get only temp from temp sensor 1
+# 2 Get only temp from temp sensor 2
+TEMP_BATTERY = int(config["DEFAULT"]["TEMP_BATTERY"])
+
+# Temperature sensor 1 name
+TEMP_1_NAME = config["DEFAULT"]["TEMP_1_NAME"]
+
+# Temperature sensor 2 name
+TEMP_2_NAME = config["DEFAULT"]["TEMP_2_NAME"]
+
+
+# --------- BMS specific settings ---------
+
+# -- LltJbd settings
+# SoC low levels
+# NOTE: SOC_LOW_WARNING is also used to calculate the Time-To-Go even if you are not using a LltJbd BMS
 SOC_LOW_WARNING = float(config["DEFAULT"]["SOC_LOW_WARNING"])
 SOC_LOW_ALARM = float(config["DEFAULT"]["SOC_LOW_ALARM"])
 
-# Daly settings
+# -- Daly settings
 # Battery capacity (amps) if the BMS does not support reading it
 BATTERY_CAPACITY = float(config["DEFAULT"]["BATTERY_CAPACITY"])
 # Invert Battery Current. Default non-inverted. Set to -1 to invert
 INVERT_CURRENT_MEASUREMENT = int(config["DEFAULT"]["INVERT_CURRENT_MEASUREMENT"])
 
-# TIME TO SOC settings [Valid values 0-100, but I don't recommend more that 20 intervals]
-# Set of SoC percentages to report on dbus. The more you specify the more it will impact system performance.
-# TIME_TO_SOC_POINTS = [100, 95, 90, 85, 80, 75, 70, 65, 60, 55, 50, 45, 40, 35, 30, 25, 20, 15, 10, 5, 0]
-# Every 5% SoC
-# TIME_TO_SOC_POINTS = [100, 95, 90, 85, 75, 50, 25, 20, 10, 0]
-TIME_TO_SOC_POINTS = _get_list_from_config("DEFAULT", "TIME_TO_SOC_POINTS")
-# Specify TimeToSoc value type: [Valid values 1,2,3]
-# TIME_TO_SOC_VALUE_TYPE = 1      # Seconds
-# TIME_TO_SOC_VALUE_TYPE = 2      # Time string HH:MN:SC
-TIME_TO_SOC_VALUE_TYPE = int(config["DEFAULT"]["TIME_TO_SOC_VALUE_TYPE"])
-# Specify how many loop cycles between each TimeToSoc updates
-TIME_TO_SOC_LOOP_CYCLES = int(config["DEFAULT"]["TIME_TO_SOC_LOOP_CYCLES"])
-# Include TimeToSoC points when moving away from the SoC point. [Valid values True,False]
-# These will be as negative time. Disabling this improves performance slightly.
-TIME_TO_SOC_INC_FROM = "True" == config["DEFAULT"]["TIME_TO_SOC_INC_FROM"]
-
-
-# Select the format of cell data presented on dbus. [Valid values 0,1,2,3]
-# 0 Do not publish all the cells (only the min/max cell data as used by the default GX)
-# 1 Format: /Voltages/Cell# (also available for display on Remote Console)
-# 2 Format: /Cell/#/Volts
-# 3 Both formats 1 and 2
-BATTERY_CELL_DATA_FORMAT = int(config["DEFAULT"]["BATTERY_CELL_DATA_FORMAT"])
-
-# Settings for ESC GreenMeter and Lipro devices
+# -- ESC GreenMeter and Lipro device settings
 GREENMETER_ADDRESS = int(config["DEFAULT"]["GREENMETER_ADDRESS"])
 LIPRO_START_ADDRESS = int(config["DEFAULT"]["LIPRO_START_ADDRESS"])
 LIPRO_END_ADDRESS = int(config["DEFAULT"]["LIPRO_END_ADDRESS"])
 LIPRO_CELL_COUNT = int(config["DEFAULT"]["LIPRO_CELL_COUNT"])
 
-PUBLISH_CONFIG_VALUES = int(config["DEFAULT"]["PUBLISH_CONFIG_VALUES"])
 
-BMS_TYPE = config["DEFAULT"]["BMS_TYPE"]
-
-
+# --------- Functions ---------
 def constrain(val, min_val, max_val):
     if min_val > max_val:
         min_val, max_val = max_val, min_val
